requirements.txt
albumentations==2.0.8
insightface==0.7.3
matplotlib==3.10.0
matplotlib-inline==0.1.6
onnx==1.18.0
onnxruntime-gpu==1.22.0
pandas==2.2.3
pillow==10.2.0
pytorch-metric-learning==2.8.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
timm==1.0.15
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchmetrics==1.7.3
torchvision==0.22.1+cu118
tornado==6.4
tqdm==4.67.1